+++ src/App.js
@@ -2,22 +2,22 @@ import './App.css';
 import { BrowserRouter, Routes, Route, Router } from 'react-router-dom';
 import Loginpage from './Components/Loginpage';
 import Homepage from './Components/Homepage';
-import Historypapers from './Papers/Historypapers';
+import Historypapers from './Papers/Board/Maharashtra/10th/Historypapers';
 import AboutUs from './Components/AboutUs';
 import Contentscreen from './Components/Contentscreen';
 import Fetchdata from './Components/Fetchdata';
 import Blogs from './Components/Blogs';
-import Englishpapers from '../src/Papers/Englishpapers'
-import HindiHalfpapers from '../src/Papers/HindiHalfpapers';
-import HindiFullpapers from '../src/Papers/HindiFullpapers';
-import SanskritFullpapers from '../src/Papers/SanskritFullpapers';
-import SanskritHalfpapers from '../src/Papers/SanskritHalfpapers';
-import Marathipapers from '../src/Papers/Marathipapers';
-import Geographypapers from '../src/Papers/Geographypapers';
-import Math1papers from '../src/Papers/Math1papers';
-import Math2papers from '../src/Papers/Math2papers';
-import Science1 from '../src/Papers/Science1';
-import Science2 from '../src/Papers/Science2';
+import Englishpapers from '../src/Papers/Board/Maharashtra/10th/Englishpapers'
+import HindiHalfpapers from '../src/Papers/Board/Maharashtra/10th/HindiHalfpapers';
+import HindiFullpapers from '../src/Papers/Board/Maharashtra/10th/HindiFullpapers';
+import SanskritFullpapers from '../src/Papers/Board/Maharashtra/10th/SanskritFullpapers';
+import SanskritHalfpapers from '../src/Papers/Board/Maharashtra/10th/SanskritHalfpapers';
+import Marathipapers from '../src/Papers/Board/Maharashtra/10th/Marathipapers';
+import Geographypapers from '../src/Papers/Board/Maharashtra/10th/Geographypapers';
+import Math1papers from '../src/Papers/Board/Maharashtra/10th/Math1papers';
+import Math2papers from '../src/Papers/Board/Maharashtra/10th/Math2papers';
+import Science1 from '../src/Papers/Board/Maharashtra/10th/Science1';
+import Science2 from '../src/Papers/Board/Maharashtra/10th/Science2';
 import Blog1 from './Components/Blogs/Blog1';
 import Blog2 from './Components/Blogs/Blog2';
 import Blog3 from './Components/Blogs/Blog3';
