+++ src/Papers/Board/Maharashtra/10th/Math1papers.js
@@ -1,28 +1,25 @@
 import React from "react";
-import Navbar from "../Components/Navbar";
-import Footer from "../Components/Footer";
-import {
-  BreadcrumbPages,
-} from "../Components/Breadcrumbs";
-import Joiningoptions from "../Components/Joiningoptions";
-import Exploremore from "../Components/Exploremore";
+import Navbar from "../../../../Components/Navbar.js";
+import Footer from "../../../../Components/Footer.js";
+import { BreadcrumbPages } from "../../../../Components/Breadcrumbs.js";
+import Joiningoptions from "../../../../Components/Joiningoptions.js";
+import Exploremore from "../../../../Components/Exploremore.js";
 import { useState, useEffect } from "react";
-import Contentscreen from "../Components/Contentscreen.js";
-import Loader from "../Components/Loader.js";
-import { sleep, topFunction } from "../Global.js";
-import StaticTag from "../Components/StaticTag.js";
-import LoginComponent from "../Components/LoginComponent.js";
+import Contentscreen from "../../../../Components/Contentscreen.js";
+import Loader from "../../../../Components/Loader.js";
+import { sleep, topFunction } from "../../../../Global.js";
+import StaticTag from "../../../../Components/StaticTag.js";
+import LoginComponent from "../../../../Components/LoginComponent.js";
 import { Helmet } from 'react-helmet';
 
-function SanskritFullPapers() {
+function Math1Papers() {
 
   const sel_med = localStorage.getItem('medium') ? localStorage.getItem('medium') : "marathi";
 
-  const pre_API = "https://padhaiplanet.com/api/v1/get-question?subject=sanskrit_full&medium=" + sel_med + "&standard=10";
+  const pre_API = "https://padhaiplanet.com/api/v1/get-question?subject=math_1&medium=" + sel_med + "&standard=10";
 
   //For fetching data
   const API_hisory_paper = pre_API;
-
 
   //API data hooks
   const [que_data, setQue_data] = useState([]);
@@ -37,7 +34,7 @@ function SanskritFullPapers() {
       await sleep(3000);
       setData(data.data);
       //⚠️⚠️⚠️ Do not remove below line!!⚠️⚠️⚠️
-      // localStorage.setItem("data_sanskrit_full_" + localStorage.getItem('medium'), JSON.stringify(data));
+      // localStorage.setItem("data_math_1_" + localStorage.getItem('medium'), JSON.stringify(data));
       document.getElementById("loader").classList.add("hidden");
       document.getElementById("parent").classList.remove("hidden");
       document.getElementById("explore").classList.remove("hidden");
@@ -47,24 +44,23 @@ function SanskritFullPapers() {
     }
   };
 
-
   useEffect(() => {
     topFunction();
     fetchdata(API_hisory_paper);
 
     //⚠️⚠️⚠️ Below commented code can be fixed. Do not remove!!⚠️⚠️⚠️
 
-    // const timestamp = localStorage.getItem('timestamp_sanskrit_full_' + localStorage.getItem('medium'));
-    // const data_sanskrit_full = localStorage.getItem('data_sanskrit_full_' + localStorage.getItem('medium'));
+    // const timestamp = localStorage.getItem('timestamp_math_1_' + localStorage.getItem('medium'));
+    // const data_math_1 = localStorage.getItem('data_math_1_' + localStorage.getItem('medium'));
 
-    // if (timestamp && data_sanskrit_full) {
+    // if (timestamp && data_math_1) {
 
     //   const check = (new Date()).getDate() > JSON.parse(timestamp).expDate;
 
     //   if (check) {
 
-    //     localStorage.removeItem('timestamp_sanskrit_full_' + localStorage.getItem('medium'));
-    //     localStorage.removeItem('data_sanskrit_full_' + localStorage.getItem('medium'));
+    //     localStorage.removeItem('timestamp_math_1_' + localStorage.getItem('medium'));
+    //     localStorage.removeItem('data_math_1_' + localStorage.getItem('medium'));
 
     //     //Adding timestamp
     //     const date = new Date().setDate(new Date().getDate() + 6);
@@ -72,14 +68,14 @@ function SanskritFullPapers() {
     //     // console.log(date);
     //     // console.log(new Date(date));
 
-    //     localStorage.setItem('timestamp_sanskrit_full_' + localStorage.getItem('medium'), JSON.stringify({
+    //     localStorage.setItem('timestamp_math_1_' + localStorage.getItem('medium'), JSON.stringify({
     //       value: "string",
     //       expDate: date,
     //     }))
 
     //     fetchdata(API_hisory_paper);
-    //   } else if(localStorage.getItem('data_sanskrit_full_' + localStorage.getItem('medium'))) {
-    //     const object = JSON.parse(localStorage.getItem('data_sanskrit_full_' + localStorage.getItem('medium')))
+    //   } else if(localStorage.getItem('data_math_1_' + localStorage.getItem('medium'))) {
+    //     const object = JSON.parse(localStorage.getItem('data_math_1_' + localStorage.getItem('medium')))
     //     setData(object.data)
     //     document.getElementById("loader").classList.add("hidden");
     //     document.getElementById("parent").classList.remove("hidden");
@@ -95,7 +91,7 @@ function SanskritFullPapers() {
     //   // console.log(date);
     //   // console.log(new Date(date));
 
-    //   localStorage.setItem('timestamp_sanskrit_full_' + localStorage.getItem('medium'), JSON.stringify({
+    //   localStorage.setItem('timestamp_math_1_' + localStorage.getItem('medium'), JSON.stringify({
     //     value: "string",
     //     expDate: date,
     //   }))
@@ -109,7 +105,6 @@ function SanskritFullPapers() {
   for (var j = 0; j < data["length"]; j++) {
     data_imp.push(data[j]);
   }
-
 
   function HandleClick(paper_no, year) {
     topFunction();
@@ -137,11 +132,11 @@ function SanskritFullPapers() {
     <div className="relative z-0">
       <Helmet>
         <title>Discover English Medium, Marathi Medium, and Sem-English Medium Maharashtra 10th SSC Board History Question Paper for 2018,2019,2020,2022,2023.</title>
-        <link rel="canonical" href="https://padhaiplanet.com/maharashtra-board-sanskrit-full-question-papers-10th-ssc" />
-        <meta name="description" content="Prepare for your exam with our Sanskrit Full question paper for the Maharashtra 10th SSC board. Prepare with the best resources and boost your chances of success in Maharashtra 10th SSC boards by practicing the previous year's question papers for 2018,2019,2020,2022,2023." />
-        <meta name="keywords" content="10th SSC Maharashtra board question papers, Last 5 years question papers, Class 10 th Maharashtra board Sanskrit Full question paper and Solutions., Question paper pdf., 10th Standard Board exam., Maharashtra Board questions paper all classes., Maharashtra Board Question and Solutions." />
+        <link rel="canonical" href="https://padhaiplanet.com/maharashtra-board-math-1-question-papers-10th-ssc" />
+        <meta name="description" content="Prepare for your exam with our Mathematics 1 question paper for the Maharashtra 10th SSC board. Prepare with the best resources and boost your chances of success in Maharashtra 10th SSC boards by practicing the previous year's question papers for 2018,2019,2020,2022,2023." />
+        <meta name="keywords" content="10th SSC Maharashtra board question papers, Last 5 years question papers, Class 10 th Maharashtra board Mathematics 1 question paper and Solutions., Question paper pdf., 10th Standard Board exam., Maharashtra Board questions paper all classes., Maharashtra Board Question and Solutions." />
       </Helmet>
-      
+
       <div>
         <div className="sticky flex top-0 z-10">
           <Navbar />
@@ -150,7 +145,7 @@ function SanskritFullPapers() {
           <StaticTag />
         </div>
         <div className="lg:px-8 md:px-6 sm:mx-4">
-          <BreadcrumbPages sub={"Sanskrit(Full)"} />
+          <BreadcrumbPages sub={"Mathematics 1"} />
         </div>
         <div id="parent" className="relative">
           <div id="go" className=" top-0 w-full mt-[50px]">
@@ -158,7 +153,7 @@ function SanskritFullPapers() {
             <div className="pb-[5%]">
               {/* <h1 className="xl:ml-[25%] lg:ml-[80%] md:ml-[25%] ml-[30%] xl:mb-[0%] lg:mb-[0%] md:mb-[0%] mb-[50%] text-center text-3xl font-bold text-white"> */}
               <h1 className="text-center xl:text-3xl lg:text-3xl md:text-3xl sm:text-2xl text-2xl font-bold text-white">
-                Sanskrit (Full)
+                Mathematics 1
               </h1>
               {/* <div className="w-[100%] xl:ml-[0%] lg:ml-[0%] md:ml-[25%] ml-[0%]"> */}
               <div className="w-full items-center">
@@ -173,15 +168,13 @@ function SanskritFullPapers() {
                         <div className="block xl:mr-[0%] lg:mr-[3%] md:mr-[0%] mr-[0%]">
                           <div className="2xl:w-[100%] xl:w-[100%] lg:w-[100%] md:w-[100%] sm:w-[100%] w-[100%] block">
                             <button
-                              onClick={(event) =>
-                                HandleClick(index1, index)
-                              }
+                              onClick={(event) => HandleClick(index1, index)}
                               type="submit"
                               className="rounded-2xl 2xl:w-[300px] xl:w-[300px] lg:w-[280px] md:w-[250px] w-[200px] text-white h-[200px] text-lg text-center font-semibold mt-[50px] bg-gradient-to-r from-[#054569] to-[#5591A9]"
                               key={index}
                             >
                               Q Paper
-                              {/* {item.papers[index1]["name"]}  */}
+                              {/* {item.papers[index1]["name"]} */}
                             </button>
                             {item.papers[index1]["solution_url"] ? (<button
                               onClick={(event) =>
@@ -227,7 +220,7 @@ function SanskritFullPapers() {
           <Joiningoptions />
         </div>
         <div id="explore" className="hidden">
-          <Exploremore sub_name={"sanskrit_full"} />
+          <Exploremore sub_name={"math_1"} />
         </div>
         <div id="footer" className="hidden">
           <Footer />
@@ -241,4 +234,4 @@ function SanskritFullPapers() {
   );
 }
 
-export default SanskritFullPapers;
+export default Math1Papers;
