+++ src/Papers/Board/Maharashtra/10th/SanskritHalfpapers.js
@@ -1,27 +1,28 @@
 import React from "react";
-import Navbar from "../Components/Navbar";
-import Footer from "../Components/Footer";
+import Navbar from "../../../../Components/Navbar.js";
+import Footer from "../../../../Components/Footer.js";
 import {
   BreadcrumbPages,
-} from "../Components/Breadcrumbs";
-import Joiningoptions from "../Components/Joiningoptions";
-import Exploremore from "../Components/Exploremore";
+} from "../../../../Components/Breadcrumbs.js";
+import Joiningoptions from "../../../../Components/Joiningoptions.js";
+import Exploremore from "../../../../Components/Exploremore.js";
 import { useState, useEffect } from "react";
-import Contentscreen from "../Components/Contentscreen.js";
-import Loader from "../Components/Loader.js";
-import { sleep, topFunction } from "../Global.js";
-import StaticTag from "../Components/StaticTag.js";
-import LoginComponent from "../Components/LoginComponent.js";
+import Contentscreen from "../../../../Components/Contentscreen.js";
+import Loader from "../../../../Components/Loader.js";
+import { sleep, topFunction } from "../../../../Global.js";
+import StaticTag from "../../../../Components/StaticTag.js";
+import LoginComponent from "../../../../Components/LoginComponent.js";
 import { Helmet } from 'react-helmet';
 
-function HindiFullPapers() {
+function SanskritHalfPapers() {
 
   const sel_med = localStorage.getItem('medium') ? localStorage.getItem('medium') : "marathi";
 
-  const pre_API = "https://padhaiplanet.com/api/v1/get-question?subject=hindi_full&medium=" + sel_med + "&standard=10";
+  const pre_API = "https://padhaiplanet.com/api/v1/get-question?subject=sanskrit_half&medium=" + sel_med + "&standard=10";
 
   //For fetching data
   const API_hisory_paper = pre_API;
+
 
   //API data hooks
   const [que_data, setQue_data] = useState([]);
@@ -36,7 +37,7 @@ function HindiFullPapers() {
       await sleep(3000);
       setData(data.data);
       //⚠️⚠️⚠️ Do not remove below line!!⚠️⚠️⚠️
-      // localStorage.setItem("data_hindi_full_" + localStorage.getItem('medium'), JSON.stringify(data));
+      // localStorage.setItem("data_sanskrit_half_" + localStorage.getItem('medium'), JSON.stringify(data));
       document.getElementById("loader").classList.add("hidden");
       document.getElementById("parent").classList.remove("hidden");
       document.getElementById("explore").classList.remove("hidden");
@@ -53,17 +54,17 @@ function HindiFullPapers() {
 
     //⚠️⚠️⚠️ Below commented code can be fixed. Do not remove!!⚠️⚠️⚠️
 
-    // const timestamp = localStorage.getItem('timestamp_hindi_full_' + localStorage.getItem('medium'));
-    // const data_hindi_full = localStorage.getItem('data_hindi_full_' + localStorage.getItem('medium'));
+    // const timestamp = localStorage.getItem('timestamp_sanskrit_half_' + localStorage.getItem('medium'));
+    // const data_sanskrit_half = localStorage.getItem('data_sanskrit_half_' + localStorage.getItem('medium'));
 
-    // if (timestamp && data_hindi_full) {
+    // if (timestamp && data_sanskrit_half) {
 
     //   const check = (new Date()).getDate() > JSON.parse(timestamp).expDate;
 
     //   if (check) {
 
-    //     localStorage.removeItem('timestamp_hindi_full_' + localStorage.getItem('medium'));
-    //     localStorage.removeItem('data_hindi_full_' + localStorage.getItem('medium'));
+    //     localStorage.removeItem('timestamp_sanskrit_half_' + localStorage.getItem('medium'));
+    //     localStorage.removeItem('data_sanskrit_half_' + localStorage.getItem('medium'));
 
     //     //Adding timestamp
     //     const date = new Date().setDate(new Date().getDate() + 6);
@@ -71,14 +72,14 @@ function HindiFullPapers() {
     //     // console.log(date);
     //     // console.log(new Date(date));
 
-    //     localStorage.setItem('timestamp_hindi_full_' + localStorage.getItem('medium'), JSON.stringify({
+    //     localStorage.setItem('timestamp_sanskrit_half_' + localStorage.getItem('medium'), JSON.stringify({
     //       value: "string",
     //       expDate: date,
     //     }))
 
     //     fetchdata(API_hisory_paper);
-    //   } else if(localStorage.getItem('data_hindi_full_' + localStorage.getItem('medium'))) {
-    //     const object = JSON.parse(localStorage.getItem('data_hindi_full_' + localStorage.getItem('medium')))
+    //   } else if(localStorage.getItem('data_sanskrit_half_' + localStorage.getItem('medium'))) {
+    //     const object = JSON.parse(localStorage.getItem('data_sanskrit_half_' + localStorage.getItem('medium')))
     //     setData(object.data)
     //     document.getElementById("loader").classList.add("hidden");
     //     document.getElementById("parent").classList.remove("hidden");
@@ -94,7 +95,7 @@ function HindiFullPapers() {
     //   // console.log(date);
     //   // console.log(new Date(date));
 
-    //   localStorage.setItem('timestamp_hindi_full_' + localStorage.getItem('medium'), JSON.stringify({
+    //   localStorage.setItem('timestamp_sanskrit_half_' + localStorage.getItem('medium'), JSON.stringify({
     //     value: "string",
     //     expDate: date,
     //   }))
@@ -136,11 +137,11 @@ function HindiFullPapers() {
     <div className="relative z-0">
       <Helmet>
         <title>Discover English Medium, Marathi Medium, and Sem-English Medium Maharashtra 10th SSC Board History Question Paper for 2018,2019,2020,2022,2023.</title>
-        <link rel="canonical" href="https://padhaiplanet.com/maharashtra-board-hindi-full-question-papers-10th-ssc" />
-        <meta name="description" content="Prepare for your exam with our Hindi Full question paper for the Maharashtra 10th SSC board. Prepare with the best resources and boost your chances of success in Maharashtra 10th SSC boards by practicing the previous year's question papers for 2018,2019,2020,2022,2023." />
-        <meta name="keywords" content="10th SSC Maharashtra board question papers, Last 5 years question papers, Class 10 th Maharashtra board Hindi Full question paper and Solutions., Question paper pdf., 10th Standard Board exam., Maharashtra Board questions paper all classes., Maharashtra Board Question and Solutions." />
+        <link rel="canonical" href="https://padhaiplanet.com/maharashtra-board-sanskrit-half-question-papers-10th-ssc" />
+        <meta name="description" content="Prepare for your exam with our Sanskrit Half question paper for the Maharashtra 10th SSC board. Prepare with the best resources and boost your chances of success in Maharashtra 10th SSC boards by practicing the previous year's question papers for 2018,2019,2020,2022,2023." />
+        <meta name="keywords" content="10th SSC Maharashtra board question papers, Last 5 years question papers, Class 10 th Maharashtra board Sanskrit Half question paper and Solutions., Question paper pdf., 10th Standard Board exam., Maharashtra Board questions paper all classes., Maharashtra Board Question and Solutions." />
       </Helmet>
-
+      
       <div>
         <div className="sticky flex top-0 z-10">
           <Navbar />
@@ -149,7 +150,7 @@ function HindiFullPapers() {
           <StaticTag />
         </div>
         <div className="lg:px-8 md:px-6 sm:mx-4">
-          <BreadcrumbPages sub={"Hindi(Full)"} />
+          <BreadcrumbPages sub={"Sanskrit(Half)"} />
         </div>
         <div id="parent" className="relative">
           <div id="go" className=" top-0 w-full mt-[50px]">
@@ -157,7 +158,7 @@ function HindiFullPapers() {
             <div className="pb-[5%]">
               {/* <h1 className="xl:ml-[25%] lg:ml-[80%] md:ml-[25%] ml-[30%] xl:mb-[0%] lg:mb-[0%] md:mb-[0%] mb-[50%] text-center text-3xl font-bold text-white"> */}
               <h1 className="text-center xl:text-3xl lg:text-3xl md:text-3xl sm:text-2xl text-2xl font-bold text-white">
-                Hindi (Full)
+                Sanskrit (Half)
               </h1>
               {/* <div className="w-[100%] xl:ml-[0%] lg:ml-[0%] md:ml-[25%] ml-[0%]"> */}
               <div className="w-full items-center">
@@ -226,7 +227,7 @@ function HindiFullPapers() {
           <Joiningoptions />
         </div>
         <div id="explore" className="hidden">
-          <Exploremore sub_name={"hindi_full"} />
+          <Exploremore sub_name={"sanskrit_half"} />
         </div>
         <div id="footer" className="hidden">
           <Footer />
@@ -240,4 +241,4 @@ function HindiFullPapers() {
   );
 }
 
-export default HindiFullPapers;
+export default SanskritHalfPapers;
